refactor(scripts): move workflow gates into scripts/workflows/ (P1)
Phase P1 of the scripts/ reorganization (issue 1877). Moves the nine files
of the workflow trust and hygiene cluster into scripts/workflows/ and updates
every caller in the same PR.

The cluster has no package.json alias and no entry in the pinned alias map,
so every caller is a literal `node scripts/...` string in a workflow, a doc,
or the quality gate. Basenames stay byte-identical, as ADR 0064 requires.

check-autofix-ci-trust.mjs also splits at the annotation-comment scoping
seam. The raw-source pass that finds `# autofix-ci-trust:` comments and
attributes them to jobs moves to scripts/workflows/autofix-trust-annotations.mjs,
leaving the checker with analysis of the parsed document alone. Behaviour is
unchanged; the existing suite already covers that region.

Swept: ci.yml run lines and comments, trunk.yml, platform-settings-drift.yml,
five workflows whose autofix-ci-trust annotations name the checker by path,
five docs, terraform/github-actions-permissions.tf, and five focused arms in
agent-quality-gate.sh. The gate test gains routing assertions for the trust
checker and the split module; the trust arm had none.

Two path derivations inside the moved tests were cwd- or depth-dependent and
are fixed: the pin test now resolves the script it spawns from import.meta.url,
and the trust test walks two directories up to reach ci.yml.
