+++ .github/workflows/claude.yml
@@ -26,7 +26,7 @@ jobs:
     # OWNER/MEMBER authors (issue_comment / review events) — never
     # automatically on a pull_request, so a machine-authored autofix PR
     # cannot reach its secret without a maintainer explicitly invoking it.
-    # See scripts/check-autofix-ci-trust.mjs.
+    # See scripts/workflows/check-autofix-ci-trust.mjs.
     if: |
       (
         github.event_name == 'issue_comment' &&
