+++ scripts/workflows/check-github-action-pins.mjs
