+++ scripts/workflows/check-notifier-coverage-helpers.mjs
