+++ scripts/workflows/check-workflow-permissions-drift.test.mjs
